Update Governance.md
Made a couple of changes in spelling and updated the link to ANSI’s Essential Requirements for Due Process.

Also @orbmis , the text below from Section 3: 

These requirements apply to activities related to the development of consensus for approval, revision, reaffirmation, and withdrawal of the "agreements data standard specification".

What would agreements data standard specification mean? Is it defined? Or more of a general concept?